refactor: collider, enemyGroup, shooting and gameScnee files
